==> 4-D115_MAGMA_summary.Rmd <==
---
title: "2018 Lynn Canal MAGMA Summary"
date: "01/29/2019"
editor_options:
  chunk_output_type: inline
output:
  html_document:
    df_print: paged
subtitle: 
creator: Chase Jalbert
---

<b> Project Overview </b> 

This is a post-season analysis of Lynn Canal D115 fishery. I want an age-specific stock composition for all major contributing age classes (>5%; 1.2, 1.3, 2.2, 2.3, other) using mark- and age-enhanced genetic mixed-stock analysis (MAGMA).    

For the purposes of this project, I am just concerned about the breakdown by age, across all stat-weeks, and stat-areas. I will assign <b>all</b> stat areas as 115-<i>00</i>, since this doesn't exist and is currently how stat area is recorded.    
 
Total season estimates are to be provided by age group. The algorithm will be run for 40 000 repetitions, and the first 20 000 repetitions will be discarded.    

Deliverable: 
    
    Point estimates, Credibility intervals
    
<b> Script Overview </b>

Now that everything is setup and verified, the model is ready. Briefly, we run 40K repetitions, burn the first 20k, and thin by 10, with 5 chains. If there are issues (e.g., no convergence) then we bump up to 80K, 40k, 100, 5, respectively. Here, I have provided both options, so commencement whatever you need.   

At this point, I will switch the scripts over to our genetics server and make use of some extra processing power.    

<b> Begin Summary Script </b>

This script takes the MAGMA output and creates summary files along with a summary work-space (R Data). Essentially it works through each chain and compiles necessary statistics presenting them into a consolidated format.

```{r setup, include = FALSE}
# This forces the working directory for ALL chunks in the notebook. Note that it must be run first, in the "setup" chunk or it will fail. 
wd <-
  "V:/Analysis/1_SEAK/Sockeye/Mixture/Lynn Canal Postseason/2018/MAGMA/"

knitr::opts_knit$set(root.dir = wd)
```

```{r libraries, message=FALSE}
if (!require("pacman")) install.packages("pacman"); library(pacman) # install pacman, if not installed

p_load(tidyverse) # use pacman to load or install+load necessary packages
```

```{r arguments}
args <-
  dget("output/MAGMA_arguments.txt")  # This will get a list with the arguments used when magma.R was run, need to set wd first

invisible(sapply(names(args)[1:4], function(arg) {
  assign(x = arg, value = args[[arg]], pos = 1)
}))  # This will load "nreps", "burn", "thin", and "nchains" to the workspace

nreps

burn

thin

nchains

alpha <- 0.10

zero <-
  "halfafish" #  This uses half a fish as the threshold for zero (based on harvest). Otherwise can be set as a numeric e.g. zero = 5e-7.

#  age_class <- new_age_class # You can re-specify this argument if u want to summarise ages withe respect to a different "age_class" vector.

groups <-
  read.table("data/groups.txt", header = TRUE, row.names = 1)

group_names <- read.table("data/group_names.txt", header = TRUE)
```

#### Compile and Create Work-space ####

```{r magma_summary}
startTime <- Sys.time()


packages <- c("reshape", "coda", "readr")

invisible( sapply( packages, function(pkg)
  while (!require( pkg, character.only = TRUE))
    install.packages(
      pkg,
      repos = c("http://rstudio.org/_packages", "http://cran.rstudio.com")
    )))

colsum <-
  function(x, group, reorder = TRUE, ...)
    t(rowsum(t(x), group = group, reorder = reorder, ...))

load("analysis/magma_data.RData")

# print(magma_data_names)

#  c("harvest", "wildpops", "hatcheries", "age_classes", "age_class", "years2run", "districts", "subdistricts", "stat_weeks", "K", "H", "A", "C", "T", "D", "S", "W")

harvest = sort_df(merge(harvest, Reduce(rbind, lapply(seq(
  D
), function(d) {
  grd = expand.grid(
    YEAR = seq(T),
    DISTRICT = d,
    SUBDISTRICT = seq(S[d]),
    STAT_WEEK = seq(W),
    stringsAsFactors = FALSE
  )
  sort_df(grd, vars = c("SUBDISTRICT", "STAT_WEEK"))
})), all = TRUE))

h <-
  tapply(harvest[, ncol(harvest)], harvest[,-ncol(harvest)], c)

n <-
  tapply(metadat$i, metadat[,-c(ncol(metadat) - 1, ncol(metadat))], length)

n[is.na(n)] <- h[is.na(h)] <- 0

paths <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      lapply(seq(S[d]), function(s) {
        lapply(seq(W), function(w) {
          paste0("output/",
                 years2run[t],
                 "/D",
                 districts[d],
                 "/S",
                 subdistricts[[d]][s],
                 "/StatWeek",
                 w)
        })
      })
    })
  })



output <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      lapply(seq(S[d]), function(s) {
        lapply(seq(W), function(w) {
          lapply(seq(nchains), function(chain) {
            rowgroup <-
              factor(paste0(
                rep(group_names[groups[, d], d], times = (nreps - burn) / thin),
                rep(seq(burn + thin, nreps, by = thin), each = K + H)
              ),
              
              levels = paste0(group_names[, d][!is.na(group_names[, d])], rep(
                seq(burn + thin, nreps, by = thin),
                each = length(group_names[, d][!is.na(group_names[, d])])
              )))
            
            file <-
              paste0(paths[[t]][[d]][[s]][[w]], "/PiP", t, d, s, w, "_Chain", chain, ".csv")
            
            
            (n[t, d, s, w] > 0) * h[t, d, s, w] * rowsum(
              colsum(
                suppressMessages(read_csv(
                  file = file, progress = FALSE
                )),
                group = factor(age_classes[age_class], levels = age_classes),
                reorder = FALSE
              ),
              group = rowgroup,
              reorder = FALSE
            )[levels(rowgroup), age_classes]
            
          })
        })
      })
    })
  })

endTime <- Sys.time()

print(endTime - startTime)


save.image("V:/Analysis/1_SEAK/Sockeye/Mixture/Lynn Canal Postseason/2018/MAGMA/output/D115summary40K.RData")
```

#### Summarize Data ####

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
###  ages & subdistricts combined by statweek

```{r R_statweek}
R <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      lapply(seq(W), function(w) {
        as.mcmc.list(lapply(seq(nchains), function(chain) {
          mcmc(Reduce("+", lapply(seq(S[d]), function(s) {
            INDEX <-
              factor(rep(group_names[!is.na(group_names[, d]), d], times = (nreps - burn) / thin), group_names[!is.na(group_names[, d]), d])
            
            xx = Reduce(cbind,
                        tapply(
                          apply(output[[t]][[d]][[s]][[w]][[chain]], 1, sum),
                          INDEX = INDEX,
                          c,
                          simplify = FALSE
                        ))
            
            dimnames(xx) <-
              list(paste0("rep", seq(burn + thin, nreps, by = thin)), levels(INDEX))
            
            xx / max(1, sum((n[t, d, , w] > 0) * h[t, d, , w]))
            
          })))
        }))
      })
    })
  })

R_statweek <-
  Reduce(rbind, lapply(seq(T), function(t) {
    Reduce(rbind, lapply(seq(D), function(d) {
      Reduce(rbind, lapply(seq(W), function(w) {
        zero <-
          ifelse(is.numeric(zero), zero, 0.5 / max(1, sum((n[t, d, , w] > 0) * h[t, d, , w])))
        
        P0 <-
          apply(apply(Reduce(rbind, R[[t]][[d]][[w]]), 2, "<", value = zero), 2, mean)
        
        smmry <-
          Reduce(cbind, summary(R[[t]][[d]][[w]], quantiles = c(alpha / 2, 1 - alpha / 2)))[, c(1, 2, 5, 6)]
        
        GR <-
          gelman.diag(R[[t]][[d]][[w]], autoburnin = FALSE, multivariate = FALSE)[[1]][, 1]
        
        group = group_names[!is.na(group_names[, d]), d]
        
        suppressWarnings(
          data.frame(
            group = group_names[!is.na(group_names[, d]), d],
            year = years2run[t],
            district = districts[d],
            subdistrict = "all",
            statweek = stat_weeks[w],
            smmry,
            GR = GR,
            P0 = P0,
            stringsAsFactors = FALSE
          )
        )
        
      }))
    }))
  }))

```

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
###  ages & statweeks combined by subdistrict

```{r R_subdistrict}
R <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      lapply(seq(S[d]), function(s) {
        as.mcmc.list(lapply(seq(nchains), function(chain) {
          mcmc(Reduce("+", lapply(seq(W), function(w) {
            INDEX <-
              factor(rep(group_names[!is.na(group_names[, d]), d], times = (nreps - burn) / thin), group_names[!is.na(group_names[, d]), d])
            
            xx = Reduce(cbind,
                        tapply(
                          apply(output[[t]][[d]][[s]][[w]][[chain]], 1, sum),
                          INDEX = INDEX,
                          c,
                          simplify = FALSE
                        ))
            
            dimnames(xx) <-
              list(paste0("rep", seq(burn + thin, nreps, by = thin)), levels(INDEX))
            
            xx / max(1, sum((n[t, d, s,] > 0) * h[t, d, s,]))
            
          })))
        }))
      })
    })
  })

R_subdistrict <-
  Reduce(rbind, lapply(seq(T), function(t) {
    Reduce(rbind, lapply(seq(D), function(d) {
      Reduce(rbind, lapply(seq(S[d]), function(s) {
        zero <-
          ifelse(is.numeric(zero), zero, 0.5 / max(1, sum((n[t, d, s,] > 0) * h[t, d, s,])))
        
        P0 <-
          apply(apply(Reduce(rbind, R[[t]][[d]][[s]]), 2, "<", value = zero), 2, mean)
        
        smmry <-
          Reduce(cbind, summary(R[[t]][[d]][[s]], quantiles = c(alpha / 2, 1 - alpha / 2)))[, c(1, 2, 5, 6)]
        
        GR <-
          gelman.diag(R[[t]][[d]][[s]], autoburnin = FALSE, multivariate = FALSE)[[1]][, 1]
        
        suppressWarnings(
          data.frame(
            group = group_names[!is.na(group_names[, d]), d],
            year = years2run[t],
            district = districts[d],
            subdistrict = subdistricts[[d]][s],
            statweek = "all",
            smmry,
            GR = GR,
            P0 = P0,
            stringsAsFactors = FALSE
          )
        )
        
      }))
    }))
  }))

```

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
### ages, subdistricts and Stat Weeks Combined by district

```{r R_district}
R <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      as.mcmc.list(lapply(seq(nchains), function(chain) {
        mcmc(Reduce("+", lapply(seq(S[d]), function(s) {
          Reduce("+", lapply(seq(W), function(w) {
            INDEX <-
              factor(rep(group_names[!is.na(group_names[, d]), d], times = (nreps - burn) / thin), group_names[!is.na(group_names[, d]), d])
            
            xx = Reduce(cbind,
                        tapply(
                          apply(output[[t]][[d]][[s]][[w]][[chain]], 1, sum),
                          INDEX = INDEX,
                          c,
                          simplify = FALSE
                        ))
            
            dimnames(xx) <-
              list(paste0("rep", seq(burn + thin, nreps, by = thin)), levels(INDEX))
            
            xx / max(1, sum((n[t, d, ,] > 0) * h[t, d, ,]))
            
          }))
        })))
      }))
    })
  })

R_district <-
  Reduce(rbind, lapply(seq(T), function(t) {
    Reduce(rbind, lapply(seq(D), function(d) {
      zero <-
        ifelse(is.numeric(zero), zero, 0.5 / max(1, sum((n[t, d, ,] > 0) * h[t, d, ,])))
      
      P0 <-
        apply(apply(Reduce(rbind, R[[t]][[d]]), 2, "<", value = zero), 2, mean)
      
      smmry <-
        Reduce(cbind, summary(R[[t]][[d]], quantiles = c(alpha / 2, 1 - alpha / 2)))[, c(1, 2, 5, 6)]
      
      GR <-
        gelman.diag(R[[t]][[d]], autoburnin = FALSE, multivariate = FALSE)[[1]][, 1]
      
      suppressWarnings(
        data.frame(
          group = group_names[!is.na(group_names[, d]), d],
          year = years2run[t],
          district = districts[d],
          subdistrict = "all",
          statweek = "all",
          smmry,
          GR = GR,
          P0 = P0,
          stringsAsFactors = FALSE
        )
      )
      
    }))
  })) 
```


  
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
###  ages combined by subdistrict & statweeks 
  
```{r R_subdistrict_statweek}
R <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      lapply(seq(S[d]), function(s) {
        lapply(seq(W), function(w) {
          as.mcmc.list(lapply(seq(nchains), function(chain) {
            INDEX <-
              factor(rep(group_names[!is.na(group_names[, d]), d], times = (nreps - burn) / thin), group_names[!is.na(group_names[, d]), d])
            
            xx = Reduce(cbind,
                        tapply(
                          apply(output[[t]][[d]][[s]][[w]][[chain]], 1, sum),
                          INDEX = INDEX,
                          c,
                          simplify = FALSE
                        ))
            
            dimnames(xx) <-
              list(paste0("rep", seq(burn + thin, nreps, by = thin)), levels(INDEX))
            
            mcmc(xx / max(1, sum((n[t, d, s, w] > 0) * h[t, d, s, w])))
            
          }))
        })
      })
    })
  })

R_subdistrict_statweek <-
  Reduce(rbind, lapply(seq(T), function(t) {
    Reduce(rbind, lapply(seq(D), function(d) {
      Reduce(rbind, lapply(seq(S[d]), function(s) {
        Reduce(rbind, lapply(seq(W), function(w) {
          zero <-
            ifelse(is.numeric(zero), zero, 0.5 / max(1, sum((n[t, d, s, w] > 0) * h[t, d, s, w])))
          
          P0 <-
            apply(apply(Reduce(rbind, R[[t]][[d]][[s]][[w]]), 2, "<", value = zero), 2, mean)
          
          smmry <-
            Reduce(cbind, summary(R[[t]][[d]][[s]][[w]], quantiles = c(alpha / 2, 1 - alpha / 2)))[, c(1, 2, 5, 6)]
          
          GR <-
            gelman.diag(R[[t]][[d]][[s]][[w]],
                        autoburnin = FALSE,
                        multivariate = FALSE)[[1]][, 1]
          
          suppressWarnings(
            data.frame(
              group = group_names[!is.na(group_names[, d]), d],
              year = years2run[t],
              district = districts[d],
              subdistrict = subdistricts[[d]][s],
              statweek = stat_weeks[w],
              smmry,
              GR = GR,
              P0 = P0,
              stringsAsFactors = FALSE
            )
          )
          
        }))
      }))
    }))
  }))
```


#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
### subdistricts and Stat Weeks Combined by district and age

```{r PiR_district}
PiR <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      as.mcmc.list(lapply(seq(nchains), function(chain) {
        mcmc(Reduce("+", lapply(seq(S[d]), function(s) {
          Reduce("+", lapply(seq(W), function(w) {
            INDEX <-
              factor(rep(group_names[!is.na(group_names[, d]), d], times = (nreps - burn) / thin), group_names[!is.na(group_names[, d]), d])
            
            xx = Reduce(cbind, by(output[[t]][[d]][[s]][[w]][[chain]], FUN = cbind, INDICES = INDEX))
            
            dimnames(xx) <-
              list(paste0("rep", seq(burn + thin, nreps, by = thin)), paste(
                rep(levels(INDEX), each = length(age_classes)),
                rep(age_classes, times = length(levels(INDEX))),
                sep = "_"
              ))
            
            xx / max(1, sum((n[t, d, ,] > 0) * h[t, d, ,]))
            
          }))
        })))
      }))
    })
  })

PiR_district <-
  Reduce(rbind, lapply(seq(T), function(t) {
    Reduce(rbind, lapply(seq(D), function(d) {
      zero <-
        ifelse(is.numeric(zero), zero, 0.5 / max(1, sum((n[t, d, ,] > 0) * h[t, d, ,])))
      
      P0 <-
        apply(apply(Reduce(rbind, PiR[[t]][[d]]), 2, "<", value = zero), 2, mean)
      
      smmry <-
        Reduce(cbind, summary(PiR[[t]][[d]], quantiles = c(alpha / 2, 1 - alpha / 2)))[, c(1, 2, 5, 6)]
      
      GR <-
        gelman.diag(PiR[[t]][[d]], autoburnin = FALSE, multivariate = FALSE)[[1]][, 1]
      
      group <-
        rep(group_names[!is.na(group_names[, d]), d], each = length(age_classes))
      
      age <-
        rep(age_classes, times = sum(!is.na(group_names[, d])))
      
      suppressWarnings(
        data.frame(
          group = group,
          age = age,
          year = years2run[t],
          district = districts[d],
          subdistrict = "all",
          statweek = "all",
          smmry,
          GR = GR,
          P0 = P0,
          stringsAsFactors = FALSE
        )
      )
      
    }))
  }))
```

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
### Stat Weeks Combined by subdistrict and age

```{r PiR_subdistrict}
PiR <-
  lapply(seq(T), function(t) {
    lapply(seq(D), function(d) {
      lapply(seq(S[d]), function(s) {
        as.mcmc.list(lapply(seq(nchains), function(chain) {
          mcmc(Reduce("+", lapply(seq(W), function(w) {
            INDEX <-
              factor(rep(group_names[!is.na(group_names[, d]), d], times = (nreps - burn) / thin), group_names[!is.na(group_names[, d]), d])
            
            xx = Reduce(cbind, by(output[[t]][[d]][[s]][[w]][[chain]], FUN = cbind, INDICES = INDEX))
            
            dimnames(xx) <-
              list(paste0("rep", seq(burn + thin, nreps, by = thin)), paste(
                rep(levels(INDEX), each = length(age_classes)),
                rep(age_classes, times = length(levels(INDEX))),
                sep = "_"
              ))
            
            xx / max(1, sum((n[t, d, s,] > 0) * h[t, d, s,]))
            
          })))
        }))
      })
    })
  })

PiR_subdistrict <-
  Reduce(rbind, lapply(seq(T), function(t) {
    Reduce(rbind, lapply(seq(D), function(d) {
      Reduce(rbind, lapply(seq(S[d]), function(s) {
        zero <-
          ifelse(is.numeric(zero), zero, 0.5 / max(1, sum((n[t, d, s,] > 0) * h[t, d, s,])))
        
        P0 <-
          apply(apply(Reduce(rbind, PiR[[t]][[d]][[s]]), 2, "<", value = zero), 2, mean)
        
        smmry <-
          Reduce(cbind, summary(PiR[[t]][[d]][[s]], quantiles = c(alpha / 2, 1 - alpha / 2)))[, c(1, 2, 5, 6)]
        
        GR <-
          gelman.diag(PiR[[t]][[d]][[s]], autoburnin = FALSE, multivariate = FALSE)[[1]][, 1]
        
        group <-
          rep(group_names[!is.na(group_names[, d]), d], each = length(age_classes))
        
        age <- rep(age_classes, times = sum(!is.na(group_names[, d])))
        
        suppressWarnings(
          data.frame(
            group = group,
            age = age,
            year = years2run[t],
            district = districts[d],
            subdistrict = subdistricts[[d]][s],
            statweek = "all",
            smmry,
            GR = GR,
            P0 = P0,
            stringsAsFactors = FALSE
          )
        )
        
      }))
    }))
  }))
```

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
### Check how the model worked using the Gelman–Rubin convergence diagnostic for multiple chains. Note that our standard is to re-run the model if GR exceeds 1.2, which sugggests non-convergence between chains.

```{r}
(
  GR_summary <- bind_cols(
    R_statweek = range(R_statweek$GR),
    R_subdistrict = range(R_subdistrict$GR),
    R_district = range(R_district$GR),
    R_subdistrict_statweek = range(R_subdistrict_statweek$GR),
    PiR_district = range(PiR_district$GR),
    PiR_subdistrict = range(PiR_subdistrict$GR)
  )
)
```

All looks good, none of the GR stats are greater than 1.2, so were in pretty good shape! I'll proceed to export a few tables in a text format. 

```{r tables}
write_tsv(
  Reduce(
    rbind,
    list(
      R_statweek = R_statweek,
      R_subdistrict = R_subdistrict,
      R_district = R_district,
      R_subdistrict_statweek = R_subdistrict_statweek
    )
  ),
  "output/magma_summariesNonagesGroups.txt",
  col_names = NA, 
  append = FALSE)

write_tsv(
  Reduce(
    rbind,
    list(PiR_subdistrict = PiR_subdistrict, PiR_district = PiR_district)
  ),
  "output/magma_summariesAgesGroups.txt",
  col_names = NA,
  append = FALSE
)
```


Now calclate sample sizes for various statistics (e.g., alleles, loci, harvest, etc.).

```{r SampleSizes}
ntab <-
  as.data.frame.table(x = tapply(metadat$i, metadat[,-c(ncol(metadat) - 1, ncol(metadat))], length),
                      responseName = "n_total")

ntab <-
  merge(x = ntab,
        y = as.data.frame.table(tapply(
          colsum(x, group = c(rep(1, sum(
            nalleles
          )), rep(2, T * C)))[, 1] != 0 &
            is.na(metadat$i), metadat[,-c(ncol(metadat) - 1, ncol(metadat))], sum
        ), responseName = "wild_genetics"))

ntab <-
  merge(x = ntab,
        y = as.data.frame.table(tapply(
          colsum(x, group = c(rep(1, sum(
            nalleles
          )), rep(2, T * C)))[, 1] == 0 &
            is.na(metadat$i), metadat[,-c(ncol(metadat) - 1, ncol(metadat))], sum
        ), responseName = "wild_no_genetics"))

ntab <-
  merge(x = ntab, y = as.data.frame.table(tapply(!is.na(metadat$i), metadat[,-c(ncol(metadat) - 1, ncol(metadat))], sum), responseName = "non_wild"))

ntab$w <- factor(ntab$w, seq(W))

harvest$STAT_WEEK <- factor(harvest$STAT_WEEK, seq(W))

colnames(ntab)[seq(ncol(harvest) - 1)] <-
  colnames(harvest)[seq(ncol(harvest) - 1)]

nhtab <- sort_df(merge(ntab, harvest))

nhtab[is.na(nhtab)] <- 0

write_csv(x = nhtab, path = "output/sample_size_&_harvest_table.csv")  # this was changed from: write.csv(nhtab, "output/sample_size_&_harvest_tableNB.csv")
# for compatibility with later scripts. I think this was a copy / paste issue on GEN3. 
```

I am going to remove everything from the workspace prior to starting the next steps, <b>careful here</b>: 
```{r cleaning}
rm(list=ls())
```


Now that the data is in a useable format and everything looks okay, I can create summary tables for reporting. The code above includes extra output which is not considered here. I decided to leave the extra code in the event that the scope of this project changes and we would like different breakdowns (e.g., by sw)

#### 2) Summary Tables ####


Summary by age across all weeks (SW25-34) and districts (D115). 
```{r age_import}
(LYNN.ages <-
   read_tsv("output/magma_summariesAgesGroups.txt")) # import the table containing ages
```


```{r age_tables}
(D115.ages <- LYNN.ages %>%
  filter(subdistrict == "0") %>%
  select(
    Age = age,
    ReportingGroup = group,
    Mean,
    SD,
    CI5 = X5.,
    CI95 = X95.
  ) %>%
  mutate(Age = ifelse(Age == "not", "Other", Age) # change the "not"'s in the age column to Other, as required in reports. 
  ) %>% 
  arrange(Age, ReportingGroup) # sort by age, then reporting groups.
)
```

A few QA checks.    

I want to see if the means fall within the CIs and make sure everything adds to 1. 

```{r final_QA}
sum(D115.ages$Mean) # 1, looks good
```


```{r export_age_final}
write_csv(
  x = D115.ages,
  path = "output/D115_age_estimates.csv",
  na = "NA",
  append = FALSE
)
```


Now, get the summary by statweek without ages. This is not given to managers but I need it to compare the rubias and magma estimates. I'll leave it here in case goals change in the future. 

```{r Nonage_import}
(LYNN <-
   read_tsv("output/magma_summariesNonAgesGroups.txt")) # import the table containing ages
```


```{r Nonage_tables}
(D115 <- LYNN %>%
  filter(subdistrict == "0") %>%
  select(
    District = district,
    StatWeek = statweek,
    ReportingGroup = group,
    Mean,
    SD,
    CI5 = X5.,
    CI95 = X95.
  ) %>% 
  arrange(StatWeek, ReportingGroup) # sort by age, then reporting groups.
)
```

A few QA checks.    

I want to see if the means fall within the CIs and make sure everything adds to 1. 

```{r Nonage_final_QA}
sum(D115$Mean) # 1, looks good
```


```{r export_Nonage_final}
write_csv(
  x = exceedCI,
  path = "output/D115_by_sw.csv",
  na = "NA",
  append = FALSE
)
```
